# public/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 3.2.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List Jobs
      description: Returns all backup jobs with their source, destination, and encryption profile details.
      operationId: listJobs
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Jobs]
      summary: Create Job
      operationId: createJob
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          description: Job with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /jobs/{id}:
    put:
      tags: [Jobs]
      summary: Update Job
      operationId: updateJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          description: Another job with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Jobs]
      summary: Delete Job
      operationId: deleteJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    Job:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Daily MySQL Backup"
        schedule:
          type: string
          example: "0 2 * * *"
        enabled:
          type: boolean
        sourceId:
          type: string
        compression:
          type: string
          enum: [NONE, GZIP, BROTLI]
        encryptionProfileId:
          type: [string, "null"]
        databases:
          type: string
          description: JSON array of database names
          example: '["mydb", "analytics"]'
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/JobDestination"
        source:
          type: object
          properties:
            name:
              type: string
            type:
              type: string
        createdAt:
          type: string
          format: date-time
    JobDestination:
      type: object
      properties:
        id:
          type: string
        configId:
          type: string
        priority:
          type: integer
        retention:
          type: string
          description: JSON retention policy
        config:
          type: object
          properties:
            name:
              type: string
            adapterId:
              type: string
    CreateJobRequest:
      type: object
      required: [name, schedule, sourceId, destinations]
      properties:
        name:
          type: string
          example: "Daily MySQL Backup"
        schedule:
          type: string
          description: Cron expression
          example: "0 2 * * *"
        sourceId:
          type: string
          description: Database source adapter ID
        destinations:
          type: array
          minItems: 1
          items:
            type: object
            required: [configId]
            properties:
              configId:
                type: string
                description: Storage destination adapter ID
              priority:
                type: integer
                default: 0
              retention:
                type: object
                description: Retention policy for this destination
        enabled:
          type: boolean
          default: true
        compression:
          type: string
          enum: [NONE, GZIP, BROTLI]
          default: GZIP
        encryptionProfileId:
          type: [string, "null"]
        notificationIds:
          type: array
          items:
            type: string
          description: Notification channel IDs
        notificationEvents:
          type: object
          description: Which events trigger notifications
        databases:
          type: array
          items:
            type: string
          description: "Specific databases to back up (empty = all)"
    UpdateJobRequest:
      type: object
      properties:
        name:
          type: string
        schedule:
          type: string
        sourceId:
          type: string
        destinations:
          type: array
          items:
            type: object
            required: [configId]
            properties:
              configId:
                type: string
              priority:
                type: integer
              retention:
                type: object
        enabled:
          type: boolean
        compression:
          type: string
          enum: [NONE, GZIP, BROTLI]
        encryptionProfileId:
          type: [string, "null"]
        notificationIds:
          type: array
          items:
            type: string
        notificationEvents:
          type: object
        databases:
          type: array
          items:
            type: string

    # ─── Executions ──────────────────────────────────────────────
...